// src/features/asistencia/types.ts
export interface CambioDiaFormValues {
  rut: string;
  nombre: string;
  terminal_code: TerminalCode;
  cabezal: string;
  date: string;
  prog_start: string;
  prog_end: string;
  reprogram_start: string;
  reprogram_end: string;
  day_off_date: string;
  day_off_start: string;
  day_off_end: string;
  day_on_date: string;
  day_on_start: string;
  day_on_end: string;
  document: File | null;
}